extract reducer from use_search
